INFIRMERIE: change modal to page for registering, editing and delete
