+++ assets/infirmerie/infirmerieEntry.vue
@@ -31,7 +31,7 @@
                     </b-col>
                     <b-col sm="2">
                         <div class="text-right">
-                            <b-btn variant="light" size="sm" @click="editEntry"
+                            <b-btn variant="light" size="sm" :to="'/edit/' + rowData.id + '/0'"
                             class="card-link"><icon scale="1.3" name="edit" color="green" class="align-text-bottom"></icon></b-btn>
                             <b-btn variant="light" size="sm" @click="deleteEntry"
                             class="card-link"><icon scale="1.3" name="trash" color="red" class="align-text-bottom"></icon></b-btn>
@@ -42,7 +42,7 @@
                     <b-col md="2" class="current-data"><strong>Arrivée :</strong> {{ arrival }}</b-col>
                     <b-col md="2" class="current-data">
                         <span v-if="rowData.datetime_sortie"><strong>Départ :</strong>{{ departure }}</span>
-                        <b-btn v-else class="mb-1" size="sm" @click="sortie">Encoder départ</b-btn>
+                        <b-btn v-else class="mb-1" size="sm" :to="'/edit/' + rowData.id + '/1'">Encoder départ</b-btn>
                     </b-col>
                     <b-col class="current-data">{{ rowData.motifs_admission }}</b-col>
                     <b-col class="current-data">{{ rowData.remarques_sortie }}</b-col>
